+++ src/ORMMapViewer/Model/Data/TangramData.cs
@@ -28,9 +28,9 @@ namespace ORMMap.Model.Data
 		}
 
 		// mapZoomLevels from 1 to 16 inclusive
-		public override double ConvertToMapZoom(double zoom)
+		public override int ConvertToMapZoom(double zoom)
 		{
-			return Math.Min(16, Math.Floor(zoom));
+			return (int) Math.Min(16, Math.Floor(zoom));
 		}
 
 		// {z}/{x}/{y}
